+++ tests/snapshot/mv3-contentScript-function/sw.js
@@ -1,0 +1,5 @@
+try {
+  importScripts("background-runtime.js", "test_txt-util_js.js", "background.js");
+} catch (e) {
+  console.error(e);
+}