+++ tests/snapshot/mv3-contentScript-string/cs.js
@@ -1,0 +1,7 @@
+;(async () => {
+	const getURL = typeof browser === "object" ? browser.runtime.getURL : chrome.runtime.getURL;
+	await import(getURL("runtime.js"));
+	await import(getURL("test_txt-util_js.js"));
+	await import(getURL("content.js"));
+})();
+null;